+++ src/renderer/viz/expressions/basic/List.js
@@ -3,6 +3,7 @@ import { checkExpression, implicitCast, getOrdinalFromIndex, checkMaxArguments }
 import ListImage from '../ListImage';
 import ListGeneric from './ListGeneric';
 import ListTransform from '../ListTransform';
+import CartoValidationError, { CartoValidationTypes as cvt } from '../../../../errors/carto-validation-error';
 
 const SUPPORTED_CHILD_TYPES = ['number', 'category', 'color', 'time', 'image', 'transformation'];
 
@@ -37,7 +38,7 @@ export default class List extends Base {
         checkMaxArguments(arguments, 1, 'list');
 
         if (!elems) {
-            throw new Error('list(): invalid parameters: must receive at least one argument');
+            throw new CartoValidationError(`${cvt.MISSING_REQUIRED} list(): invalid parameters: must receive at least one argument.`);
         }
 
         if (!Array.isArray(elems)) {
@@ -47,7 +48,7 @@ export default class List extends Base {
         elems = elems.map(implicitCast);
 
         if (!elems.length) {
-            throw new Error('list(): invalid parameters: must receive at least one argument');
+            throw new CartoValidationError(`${cvt.MISSING_REQUIRED} list(): invalid parameters: must receive at least one argument.`);
         }
 
         elems.map((item, index) => {
@@ -63,14 +64,14 @@ export default class List extends Base {
         this._setTypes();
 
         if (SUPPORTED_CHILD_TYPES.indexOf(this.childType) === -1) {
-            throw new Error(`list(): invalid parameters type: ${this.childType}`);
+            throw new CartoValidationError(`${cvt.INCORRECT_TYPE} list(): invalid parameters type: ${this.childType}.`);
         }
 
         this.elems.map((item, index) => {
             checkExpression('list', `item[${index}]`, index, item);
 
             if (item.type !== this.childType) {
-                throw new Error(`list(): invalid ${getOrdinalFromIndex(index + 1)} parameter type, invalid argument type combination`);
+                throw new CartoValidationError(`${cvt.INCORRECT_TYPE} list(): invalid ${getOrdinalFromIndex(index + 1)} parameter type, invalid argument type combination.`);
             }
         });