+++ test/unit/source/MVT.test.js
@@ -45,7 +45,7 @@ describe('sources/MVT', () => {
             const worker = new MVTWorker();
             expect(() => {
                 worker.decodeProperty(metadata, 'wadus', 'this is not a number');
-            }).toThrowError(/MVT decoding error. Metadata property \'wadus\' is of type \'number\' but the MVT tile contained a feature property of type string: \'this is not a number\'/);
+            }).toThrowError(/MVT decoding error. Metadata property \'wadus\' is of type \'number\' but the MVT tile contained a feature property of type 'string': \'this is not a number\'/);
         });
         it('should throw an error when the property type is number and the metadata declared type is category', () => {
             const metadata = {
@@ -59,7 +59,7 @@ describe('sources/MVT', () => {
             const worker = new MVTWorker();
             expect(() => {
                 worker.decodeProperty(metadata, 'wadus', 123);
-            }).toThrowError(/MVT decoding error. Metadata property \'wadus\' is of type \'category\' but the MVT tile contained a feature property of type number: \'123\'/);
+            }).toThrowError(/MVT decoding error. Metadata property \'wadus\' is of type \'category\' but the MVT tile contained a feature property of type 'number': \'123\'/);
         });
     });
 });